Refactoring the main browser window out of the index.js
