+++ app/mainAppWindow/index.js
@@ -1,0 +1,107 @@
+
+const { shell, BrowserWindow } = require('electron');
+const windowStateKeeper = require('electron-window-state');
+const path = require('path');
+const login = require('../login');
+const iconPath = path.join(__dirname, '..', 'assets', 'icons', 'icon-96x96.png');
+const customCSS = require('../customCSS');
+const Menus = require('../menus');
+const notifications = require('../notifications');
+const onlineOffline = require('../onlineOffline');
+
+let window = null;
+
+exports.onAppReady = function onAppReady() {
+	window = createWindow();
+	new Menus(window, config, iconPath);
+
+	window.on('page-title-updated', (event, title) => {
+		window.webContents.send('page-title', title);
+	});
+
+	if (config.enableDesktopNotificationsHack) {
+		notifications.addDesktopNotificationHack(iconPath);
+	}
+
+	window.webContents.on('new-window', onNewWindow);
+
+	login.handleLoginDialogTry(window);
+	if (config.onlineOfflineReload) {
+		onlineOffline.reloadPageWhenOfflineToOnline(window, config.url);
+	}
+
+	window.webContents.setUserAgent(config.chromeUserAgent);
+
+	if (!config.minimized) {
+		window.once('ready-to-show', () => window.show());
+	}
+
+	window.webContents.on('did-finish-load', () => {
+		customCSS.onDidFinishLoad(window.webContents)
+	});
+
+	window.on('closed', () => { window = null; });
+
+	window.loadURL(config.url);
+
+	if (config.webDebug) {
+		window.openDevTools();
+	}
+}
+
+exports.onAppSecondInstance = function onAppSecondInstance (args) {
+	console.log('second-instance started');
+	if (window) {
+		console.log('focusing on window');
+		if (window.isMinimized()) window.restore();
+		window.focus();
+		window.webContents.send('openUrl', args[0]);
+	}
+}
+
+function onNewWindow(event, url, frame, disposition) {
+	if (url.startsWith('https://teams.microsoft.com/l/meetup-join')) {
+		event.preventDefault();
+		window.loadURL(url);
+	} else if ((disposition !== 'background-tab') && (url !== 'about:blank')) {
+		event.preventDefault();
+		shell.openExternal(url);
+	}
+};
+
+function createWindow() {
+	// Load the previous state with fallback to defaults
+	const windowState = windowStateKeeper({
+		defaultWidth: 0,
+		defaultHeight: 0,
+	});
+
+	// Create the window
+	const window = new BrowserWindow({
+		x: windowState.x,
+		y: windowState.y,
+
+		width: windowState.width,
+		height: windowState.height,
+		backgroundColor: '#fff',
+
+		show: false,
+		autoHideMenuBar: true,
+		icon: iconPath,
+
+		webPreferences: {
+			partition: config.partition,
+			preload: path.join(__dirname, '..', 'browser', 'index.js'),
+			nativeWindowOpen: true,
+			plugins: true,
+			nodeIntegration: false,
+		},
+	});
+
+	windowState.manage(window);
+	window.eval = global.eval = function () { // eslint-disable-line no-eval
+		throw new Error('Sorry, this app does not support window.eval().');
+	};
+
+	return window;
+}
